+++ R/plot.cidprop.R
@@ -407,6 +407,11 @@ plot.cidprop <- function(x,
                  sign = "\u2265 ")
     #
     min.cid.above.null <- min(cid.above.null, na.rm = TRUE)
+    #
+    if (avail.cid.below.null & prop.within.cid ==  0) {
+      dat.u %<>%
+        mutate(sign = if_else(Threshold == min.cid.above.null, "> ", sign))
+    }
   }
   #
   if (prop.within.cid > 0) {
@@ -422,7 +427,7 @@ plot.cidprop <- function(x,
         formatN(c(max.cid.below.null, min.cid.above.null),
                 digits = digits.cid, big.mark = big.mark)
       #
-     within.cid <- paste(">", within.cid[1], "to", "<", within.cid[2])
+     within.cid <- paste(within.cid[1], "to", within.cid[2])
     }
     else if (avail.cid.below.null) {
       within.cid <-
@@ -437,6 +442,8 @@ plot.cidprop <- function(x,
       within.cid <- paste("<", within.cid)
     }
   }
+  else
+    within.cid <- ""
   #
   dat.cid <- rbind(dat.l, dat.w, dat.u)
   #